requirements.txt
# Splatoon3 Assistant Dependencies

# HTTP client with HTTP/2 support
httpx[http2]>=0.25.0

# HTML parsing for version extraction
beautifulsoup4>=4.12.0

# PostgreSQL
psycopg2-binary>=2.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
